Historic city map | Google Javascript API
The Library and the City of Riihimäki have requested for a historic map
of Riihimäki based on the decade before WW2, (Year 1939 to be spesific).
This Map represents historic locations, buildings and stories based in
the Riihimäki region.
Info-Bubble will be used as the foundating branch to the project